LPD-50980 Rename to ImportFragments for better file usage understanding
