+++ modules/apps/fragment/fragment-web/src/main/resources/META-INF/resources/js/ImportFragments.tsx
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import {Import} from '@liferay/layout-js-components-web';
+import {ImportFragments as Import} from '@liferay/layout-js-components-web';
 import React from 'react';
 
 interface Props {
